+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ### 7.0.0
 
+##### Breaking Changes
+
+-   Deprecated prop `hideDoneBar` has been removed
+-   Deprecated prop `placeholderTextColor` has been removed
+
 ##### Chore
 
 -   Remove deprecated ColorPropType